anvil_zksync_api_decl

Trait ZksNamespaceServer

pub trait ZksNamespaceServer:
    Sized
    + Send
    + Sync
    + 'static {

    // Required methods
    fn estimate_fee<'life0, 'async_trait>(
        &'life0 self,
        req: CallRequest,
        state_override: Option<StateOverride>,
    ) -> Pin<Box<dyn Future<Output = Result<Fee, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn estimate_gas_l1_to_l2<'life0, 'async_trait>(
        &'life0 self,
        req: CallRequest,
        state_override: Option<StateOverride>,
    ) -> Pin<Box<dyn Future<Output = Result<U256, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_bridgehub_contract<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<H160>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_main_l1_contract<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<H160, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_l2_multicall3<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<H160>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_testnet_paymaster<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<H160>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_timestamp_asserter<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<H160>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_bridge_contracts<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<BridgeAddresses, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_base_token_l1_address<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<H160, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn l1_chain_id<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<U64, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_l2_to_l1_log_proof<'life0, 'async_trait>(
        &'life0 self,
        tx_hash: H256,
        index: Option<usize>,
    ) -> Pin<Box<dyn Future<Output = Result<Option<L2ToL1LogProof>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_l1_batch_number<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<U64, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_l2_block_range<'life0, 'async_trait>(
        &'life0 self,
        batch: L1BatchNumber,
    ) -> Pin<Box<dyn Future<Output = Result<Option<(U64, U64)>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_block_details<'life0, 'async_trait>(
        &'life0 self,
        block_number: L2BlockNumber,
    ) -> Pin<Box<dyn Future<Output = Result<Option<BlockDetails>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_transaction_details<'life0, 'async_trait>(
        &'life0 self,
        hash: H256,
    ) -> Pin<Box<dyn Future<Output = Result<Option<TransactionDetails>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_raw_block_transactions<'life0, 'async_trait>(
        &'life0 self,
        block_number: L2BlockNumber,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Transaction>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_l1_batch_details<'life0, 'async_trait>(
        &'life0 self,
        batch: L1BatchNumber,
    ) -> Pin<Box<dyn Future<Output = Result<Option<L1BatchDetails>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_bytecode_by_hash<'life0, 'async_trait>(
        &'life0 self,
        hash: H256,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_l1_gas_price<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<U64, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_fee_params<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<FeeParams, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_protocol_version<'life0, 'async_trait>(
        &'life0 self,
        version_id: Option<u16>,
    ) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolVersion>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_proof<'life0, 'async_trait>(
        &'life0 self,
        address: H160,
        keys: Vec<H256>,
        l1_batch_number: L1BatchNumber,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Proof>, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_batch_fee_input<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<PubdataIndependentBatchFeeModelInput, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn gas_per_pubdata<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<U256, ErrorObject<'static>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }

}

Server trait implementation for the ZksNamespace RPC API.

Required Methods

fn estimate_fee<'life0, 'async_trait>( &'life0 self, req: CallRequest, state_override: Option<StateOverride>, ) -> Pin<Box<dyn Future<Output = Result<Fee, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn estimate_gas_l1_to_l2<'life0, 'async_trait>( &'life0 self, req: CallRequest, state_override: Option<StateOverride>, ) -> Pin<Box<dyn Future<Output = Result<U256, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_bridgehub_contract<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<H160>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_main_l1_contract<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<H160, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_l2_multicall3<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<H160>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_testnet_paymaster<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<H160>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_timestamp_asserter<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<H160>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_bridge_contracts<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<BridgeAddresses, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_base_token_l1_address<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<H160, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn l1_chain_id<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<U64, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_l2_to_l1_log_proof<'life0, 'async_trait>( &'life0 self, tx_hash: H256, index: Option<usize>, ) -> Pin<Box<dyn Future<Output = Result<Option<L2ToL1LogProof>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_l1_batch_number<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<U64, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_l2_block_range<'life0, 'async_trait>( &'life0 self, batch: L1BatchNumber, ) -> Pin<Box<dyn Future<Output = Result<Option<(U64, U64)>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_block_details<'life0, 'async_trait>( &'life0 self, block_number: L2BlockNumber, ) -> Pin<Box<dyn Future<Output = Result<Option<BlockDetails>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_transaction_details<'life0, 'async_trait>( &'life0 self, hash: H256, ) -> Pin<Box<dyn Future<Output = Result<Option<TransactionDetails>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_raw_block_transactions<'life0, 'async_trait>( &'life0 self, block_number: L2BlockNumber, ) -> Pin<Box<dyn Future<Output = Result<Vec<Transaction>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_l1_batch_details<'life0, 'async_trait>( &'life0 self, batch: L1BatchNumber, ) -> Pin<Box<dyn Future<Output = Result<Option<L1BatchDetails>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_bytecode_by_hash<'life0, 'async_trait>( &'life0 self, hash: H256, ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_l1_gas_price<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<U64, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_fee_params<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<FeeParams, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_protocol_version<'life0, 'async_trait>( &'life0 self, version_id: Option<u16>, ) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolVersion>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_proof<'life0, 'async_trait>( &'life0 self, address: H160, keys: Vec<H256>, l1_batch_number: L1BatchNumber, ) -> Pin<Box<dyn Future<Output = Result<Option<Proof>, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn get_batch_fee_input<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<PubdataIndependentBatchFeeModelInput, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

fn gas_per_pubdata<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<U256, ErrorObject<'static>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.